Add functionality to dark-light-mode and changing the degrees, Fahrenheit <--> Celsius
Issue: Need a way to change both the button text and the content itself . However, if I try to pass in data, then it would be null if I click the button since there is no data passed through. If I try to do it through displayData, then it will continuously change on search.